# examples/react-todos/client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "react-todos",
    "description": "",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/todos/{id}": {
      "get": {
        "tags": [
          "todos"
        ],
        "operationId": "getApiTodosById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The todo, or null if unknown",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/TodoDetail"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "todos"
        ],
        "operationId": "deleteTodo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the todo to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "No todo with that id"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "TodoDetail": {
        "type": "object",
        "description": "Wire shape of a single-todo read. (Named apart from the list DTO in\n`../route.rs` — OpenAPI schema names are global.)",
        "required": [
          "id",
          "title",
          "done"
        ],
        "properties": {
          "done": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "title": {
            "type": "string"
          }
        }
      }
    }
  }
}